Create passthrough chain element
