+++ pkg/networkservice/common/passthrough/replacelabels/client.go
@@ -1,3 +1,5 @@
+// Copyright (c) 2021 Doc.ai and/or its affiliates.
+//
 // Copyright (c) 2022 Cisco and/or its affiliates.
 //
 // SPDX-License-Identifier: Apache-2.0
