+++ pkg/networkservice/common/passthrough/replacensename/client_test.go
@@ -14,20 +14,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package replacelabels_test
+package replacensename_test
 
 import (
 	"context"
 	"testing"
 
 	"go.uber.org/goleak"
+	"google.golang.org/grpc"
 
+	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/stretchr/testify/require"
 
 	"github.com/networkservicemesh/api/pkg/api/networkservice"
 
-	"github.com/networkservicemesh/sdk/pkg/networkservice/common/replacelabels"
+	"github.com/networkservicemesh/sdk/pkg/networkservice/common/passthrough/replacensename"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/core/chain"
+	"github.com/networkservicemesh/sdk/pkg/networkservice/core/next"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/utils/checks/checkclose"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/utils/checks/checkrequest"
 	"github.com/networkservicemesh/sdk/pkg/networkservice/utils/metadata"
@@ -36,26 +39,45 @@ import (
 func TestClient(t *testing.T) {
 	t.Cleanup(func() { goleak.VerifyNone(t) })
 
-	chainLabels := map[string]string{"1": "A", "2": "B"}
 	client := chain.NewNetworkServiceClient(
 		metadata.NewClient(),
-		replacelabels.NewClient(chainLabels),
+		replacensename.NewClient(),
+		&setNSENameClient{name: "nse-name1"},
 		checkrequest.NewClient(t, func(t *testing.T, r *networkservice.NetworkServiceRequest) {
-			require.Equal(t, chainLabels, r.Connection.Labels)
+			require.Equal(t, "nse-name1", r.Connection.NetworkServiceEndpointName)
 		}),
 		checkclose.NewClient(t, func(t *testing.T, c *networkservice.Connection) {
-			require.Equal(t, chainLabels, c.Labels)
+			require.Equal(t, "nse-name1", c.NetworkServiceEndpointName)
 		}),
 	)
-
-	// Create the request with any labels
 	req := &networkservice.NetworkServiceRequest{
-		Connection: &networkservice.Connection{Id: "nsc-1", Labels: map[string]string{"3": "C"}},
+		Connection: &networkservice.Connection{Id: "nsc-1"},
 	}
 	conn, err := client.Request(context.Background(), req)
 	require.NoError(t, err)
-	require.Equal(t, map[string]string{"3": "C"}, conn.Labels)
+
+	// Change NetworkServiceEndpointName to another name
+	conn.NetworkServiceEndpointName = "nse-name2"
+	conn, err = client.Request(context.Background(), req)
+	require.Equal(t, "nse-name2", conn.NetworkServiceEndpointName)
+	require.NoError(t, err)
 
 	_, err = client.Close(context.Background(), conn)
 	require.NoError(t, err)
+}
+
+// setNSENameClient sets NetworkServiceEndpointName only if it is empty
+type setNSENameClient struct {
+	name string
+}
+
+func (s *setNSENameClient) Request(ctx context.Context, request *networkservice.NetworkServiceRequest, opts ...grpc.CallOption) (*networkservice.Connection, error) {
+	if request.GetConnection().NetworkServiceEndpointName == "" {
+		request.GetConnection().NetworkServiceEndpointName = s.name
+	}
+	return next.Client(ctx).Request(ctx, request, opts...)
+}
+
+func (s *setNSENameClient) Close(ctx context.Context, conn *networkservice.Connection, opts ...grpc.CallOption) (*empty.Empty, error) {
+	return next.Client(ctx).Close(ctx, conn, opts...)
 }
